Simplify bank-details.py usage
fetchy.py fetches html files into the directory from where it is called. Since
no information otherwise is given, an user will probably run it in the
directory where the script is. In that case, bank-details.py won't work before
those files are moved into a specific directory. That happens because, while
bank-details.py has the logic to read those files from a specific directory,
that directory is hardcoded. To mitigate that, this patch changes the hardcoded
directory to the one where the scripts are.

While this is an enhancement to the usability of the package, one might think
to improve it in the future, for instance by taking the directory name as an
argument